Update interact.py
Some targets show a mismatch between the cross-hair and the star's position on interact_sky(). The mismatch is also present for all or most of the nearby stars.

Examples include "duplicate" or "split" IDs (e.g. TIC 150391195), high-proper stars (e.g. TIC 228760807), and other targets that do not have duplicates, splits, or high proper motion (e.g. TIC 294183478).

This is a proposed solution, where interact.py uses the TIC catalog instead of the GaiaDR2 catalog to plot stars' positions. This produces an exact match between the cross-hair and the target's position.

An alternative solution could be to use the "I/355/gaiadr3" catalog but the match is still not exact.